RS485-Server/services/api-key-manager/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
cryptography==42.0.2
python-dotenv==1.0.1 